backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
pypdf==5.4.0
python-dotenv==1.1.0
openai==1.75.0
ollama==0.4.8
